+++ Kernel/Modules/AdminCustomerUser.pm
@@ -830,14 +830,27 @@ sub _Overview {
         );
     }
 
-    my %ListAllItems = $CustomerUserObject->CustomerSearch(
-        Search => '*',
-        Limit  => 999999,
-        Valid  => 0,
-    );
+    # get config object
+    my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
 
     # same Limit as $Self->{CustomerUserMap}->{CustomerUserSearchListLimit}
-    my $Limit = 250;
+    # smallest Limit from all sources
+    my $Limit = 400;
+    SOURCE:
+    for my $Count ( '', 1 .. 10 ) {
+        next SOURCE if !$ConfigObject->Get("CustomerUser$Count");
+        my $CustomerUserMap = $ConfigObject->Get("CustomerUser$Count");
+        next SOURCE if !$CustomerUserMap->{CustomerUserSearchListLimit};
+        if ( $CustomerUserMap->{CustomerUserSearchListLimit} < $Limit ) {
+            $Limit = $CustomerUserMap->{CustomerUserSearchListLimit};
+        }
+    }
+
+    my %ListAllItems = $CustomerUserObject->CustomerSearch(
+        Search => $Param{Search},
+        Limit  => $Limit + 1,
+        Valid  => 0,
+    );
 
     if ( keys %ListAllItems <= $Limit ) {
         my $ListAllItems = keys %ListAllItems;
@@ -850,12 +863,11 @@ sub _Overview {
         );
     }
 
-    my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
-
     # when there is no data to show, a message is displayed on the table with this colspan
     my $ColSpan = 6;
 
     if ( $Param{Search} ) {
+
         my %List = $CustomerUserObject->CustomerSearch(
             Search => $Param{Search},
             Valid  => 0,
@@ -1319,7 +1331,7 @@ sub _Edit {
                                 Data => {%Param},
                             );
                             if (
-                                ref $ParamItem->{Data} eq 'HASH'
+                                ref $ParamItem->{Data}   eq 'HASH'
                                 || ref $Preference{Data} eq 'HASH'
                                 )
                             {